+++ .npmrc
@@ -1,0 +1,4 @@
+legacy-peer-deps=true
+strict-peer-deps=false
+fund=false
+loglevel=error
